+++ addOns/alertFilters/src/main/java/org/zaproxy/zap/extension/alertFilters/AlertFilterAPI.java
@@ -22,7 +22,9 @@ package org.zaproxy.zap.extension.alertFilters;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
+import java.util.Set;
 import net.sf.json.JSONObject;
+import org.apache.commons.configuration.ConfigurationException;
 import org.apache.log4j.Logger;
 import org.zaproxy.zap.extension.api.ApiAction;
 import org.zaproxy.zap.extension.api.ApiException;
@@ -44,9 +46,12 @@ public class AlertFilterAPI extends ApiImplementor {
     private static final String PREFIX = "alertFilter";
 
     private static final String VIEW_ALERT_FILTER_LIST = "alertFilterList";
+    private static final String VIEW_GLOBAL_ALERT_FILTER_LIST = "globalAlertFilterList";
 
     private static final String ACTION_ADD_ALERT_FILTER = "addAlertFilter";
     private static final String ACTION_REMOVE_ALERT_FILTER = "removeAlertFilter";
+    private static final String ACTION_ADD_GLOBAL_ALERT_FILTER = "addGlobalAlertFilter";
+    private static final String ACTION_REMOVE_GLOBAL_ALERT_FILTER = "removeGlobalAlertFilter";
 
     private static final String PARAM_CONTEXT_ID = "contextId";
     private static final String PARAM_RULE_ID = "ruleId";
@@ -76,6 +81,10 @@ public class AlertFilterAPI extends ApiImplementor {
                 new ApiView(VIEW_ALERT_FILTER_LIST, new String[] {PARAM_CONTEXT_ID});
         alertFilterList.setDescriptionTag("alertFilters.api.view.alertFilterList");
         this.addApiView(alertFilterList);
+
+        ApiView globalAlertFilterList = new ApiView(VIEW_GLOBAL_ALERT_FILTER_LIST);
+        globalAlertFilterList.setDescriptionTag("alertFilters.api.view.globalAlertFilterList");
+        this.addApiView(globalAlertFilterList);
 
         ApiAction addAlertFilter =
                 new ApiAction(
@@ -112,6 +121,43 @@ public class AlertFilterAPI extends ApiImplementor {
                         });
         removeAlertFilter.setDescriptionTag("alertFilters.api.action.removeAlertFilter");
         this.addApiAction(removeAlertFilter);
+
+        ApiAction addGlobalAlertFilter =
+                new ApiAction(
+                        ACTION_ADD_GLOBAL_ALERT_FILTER,
+                        new String[] {PARAM_RULE_ID, PARAM_NEW_LEVEL},
+                        new String[] {
+                            PARAM_URL,
+                            PARAM_URL_IS_REGEX,
+                            PARAM_PARAMETER,
+                            PARAM_ENABLED,
+                            PARAM_PARAMETER_IS_REGEX,
+                            PARAM_ATTACK,
+                            PARAM_ATTACK_IS_REGEX,
+                            PARAM_EVIDENCE,
+                            PARAM_EVIDENCE_IS_REGEX,
+                        });
+        addGlobalAlertFilter.setDescriptionTag("alertFilters.api.action.addGlobalAlertFilter");
+        this.addApiAction(addGlobalAlertFilter);
+
+        ApiAction removeGlobalAlertFilter =
+                new ApiAction(
+                        ACTION_REMOVE_GLOBAL_ALERT_FILTER,
+                        new String[] {PARAM_RULE_ID, PARAM_NEW_LEVEL},
+                        new String[] {
+                            PARAM_URL,
+                            PARAM_URL_IS_REGEX,
+                            PARAM_PARAMETER,
+                            PARAM_ENABLED,
+                            PARAM_PARAMETER_IS_REGEX,
+                            PARAM_ATTACK,
+                            PARAM_ATTACK_IS_REGEX,
+                            PARAM_EVIDENCE,
+                            PARAM_EVIDENCE_IS_REGEX,
+                        });
+        removeGlobalAlertFilter.setDescriptionTag(
+                "alertFilters.api.action.removeGlobalAlertFilter");
+        this.addApiAction(removeGlobalAlertFilter);
     }
 
     @Override
@@ -134,9 +180,18 @@ public class AlertFilterAPI extends ApiImplementor {
                                 .getAlertFilters();
 
                 for (AlertFilter af : afs) {
-                    listResponse.addItem(buildResponseFromAlertFilter(af));
+                    listResponse.addItem(buildResponseFromAlertFilter(af, true));
                 }
                 return listResponse;
+
+            case VIEW_GLOBAL_ALERT_FILTER_LIST:
+                ApiResponseList globalListResponse = new ApiResponseList(name);
+                Set<AlertFilter> gafs = extension.getParam().getGlobalAlertFilters();
+
+                for (AlertFilter af : gafs) {
+                    globalListResponse.addItem(buildResponseFromAlertFilter(af, false));
+                }
+                return globalListResponse;
 
             default:
                 throw new ApiException(ApiException.Type.BAD_VIEW);
@@ -194,6 +249,58 @@ public class AlertFilterAPI extends ApiImplementor {
 
                 return ApiResponseElement.FAIL;
 
+            case ACTION_ADD_GLOBAL_ALERT_FILTER:
+                af =
+                        new AlertFilter(
+                                -1,
+                                ApiUtils.getIntParam(params, PARAM_RULE_ID),
+                                ApiUtils.getIntParam(params, PARAM_NEW_LEVEL),
+                                ApiUtils.getOptionalStringParam(params, PARAM_URL),
+                                getParam(params, PARAM_URL_IS_REGEX, false),
+                                ApiUtils.getOptionalStringParam(params, PARAM_PARAMETER),
+                                getParam(params, PARAM_PARAMETER_IS_REGEX, false),
+                                ApiUtils.getOptionalStringParam(params, PARAM_ATTACK),
+                                getParam(params, PARAM_ATTACK_IS_REGEX, false),
+                                ApiUtils.getOptionalStringParam(params, PARAM_EVIDENCE),
+                                getParam(params, PARAM_EVIDENCE_IS_REGEX, false),
+                                getParam(params, PARAM_ENABLED, true));
+
+                // TODO more validation, esp url!
+                if (extension.getParam().addAlertFilter(af)) {
+                    try {
+                        extension.getParam().getConfig().save();
+                    } catch (ConfigurationException e) {
+                        throw new ApiException(Type.INTERNAL_ERROR, e);
+                    }
+                    return ApiResponseElement.OK;
+                }
+                return ApiResponseElement.FAIL;
+
+            case ACTION_REMOVE_GLOBAL_ALERT_FILTER:
+                af =
+                        new AlertFilter(
+                                -1,
+                                ApiUtils.getIntParam(params, PARAM_RULE_ID),
+                                ApiUtils.getIntParam(params, PARAM_NEW_LEVEL),
+                                ApiUtils.getOptionalStringParam(params, PARAM_URL),
+                                getParam(params, PARAM_URL_IS_REGEX, false),
+                                ApiUtils.getOptionalStringParam(params, PARAM_PARAMETER),
+                                getParam(params, PARAM_PARAMETER_IS_REGEX, false),
+                                ApiUtils.getOptionalStringParam(params, PARAM_ATTACK),
+                                getParam(params, PARAM_ATTACK_IS_REGEX, false),
+                                ApiUtils.getOptionalStringParam(params, PARAM_EVIDENCE),
+                                getParam(params, PARAM_EVIDENCE_IS_REGEX, false),
+                                getParam(params, PARAM_ENABLED, true));
+                if (extension.getParam().removeFilter(af)) {
+                    try {
+                        extension.getParam().getConfig().save();
+                    } catch (ConfigurationException e) {
+                        throw new ApiException(Type.INTERNAL_ERROR, e);
+                    }
+                    return ApiResponseElement.OK;
+                }
+                return ApiResponseElement.FAIL;
+
             default:
                 throw new ApiException(Type.BAD_ACTION);
         }
@@ -205,9 +312,11 @@ public class AlertFilterAPI extends ApiImplementor {
      * @param af the AlertFilter
      * @return the api response
      */
-    private ApiResponse buildResponseFromAlertFilter(AlertFilter af) {
+    private ApiResponse buildResponseFromAlertFilter(AlertFilter af, boolean includeContext) {
         Map<String, String> fields = new HashMap<>();
-        fields.put(PARAM_CONTEXT_ID, Integer.toString(af.getContextId()));
+        if (includeContext) {
+            fields.put(PARAM_CONTEXT_ID, Integer.toString(af.getContextId()));
+        }
         fields.put(PARAM_RULE_ID, Integer.toString(af.getRuleId()));
         fields.put(PARAM_NEW_LEVEL, Integer.toString(af.getNewRisk()));
         fields.put(PARAM_URL, af.getUrl());