+++ addOns/alertFilters/CHANGELOG.md
@@ -6,6 +6,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ## Unreleased
 
 - Added support for parameter regex, attack and evidence strings and regexes (Issue 5574)
+- Added support for global alert filters (Issue 5575)
 
 ## [8] - 2019-06-07